ScheduleDAO 인터페이스 추가, ScheduleMapper NAMESPACE 변경
